Add Mulitple guides

+++ deepdream/deepdream.py
@@ -1,4 +1,5 @@
-# imports and basic notebook setup
+import os
+os.environ['GLOG_minloglevel'] = '2'
 from cStringIO import StringIO
 import numpy as np
 import scipy.ndimage as nd
@@ -9,7 +10,6 @@ from google.protobuf import text_format
 import shutil
 import caffe
 from multiprocessing import Process
-import os
 
 def showarray(a, fmt='jpeg'):
     a = np.uint8(np.clip(a, 0, 255))
@@ -72,7 +72,7 @@ def make_step(net, step_size=1.5, end='inception_4c/output',
         bias = net.transformer.mean['data']
         src.data[:] = np.clip(src.data, -bias, 255-bias)
 
-def deepdream(filename, net, base_img, iter_n=10, octave_n=4, octave_scale=1.4, end='inception_4c/output', clip=True, model="", jitter=32, guide_features=None,guide=False,**step_params):
+def deepdream(filename, net, base_img, iter_n=10, octave_n=4, octave_scale=1.4, end='inception_4c/output', clip=True, model="", jitter=32, guide_features=None,guide=None,guide_file="",**step_params):
     # prepare base images for all octaves
     octaves = [preprocess(net, base_img)]
     for i in xrange(octave_n-1):
@@ -80,6 +80,8 @@ def deepdream(filename, net, base_img, iter_n=10, octave_n=4, octave_scale=1.4, 
 
     src = net.blobs['data']
     detail = np.zeros_like(octaves[-1]) # allocate image for network-produced details
+    print("OCTS:" + str(octave_n))
+
     for octave, octave_base in enumerate(octaves[::-1]):
         h, w = octave_base.shape[-2:]
         if octave > 0:
@@ -89,6 +91,7 @@ def deepdream(filename, net, base_img, iter_n=10, octave_n=4, octave_scale=1.4, 
 
         src.reshape(1,3,h,w) # resize the network's input image size
         src.data[0] = octave_base+detail
+
         for i in xrange(iter_n):
             make_step(net, end=end, jitter=jitter, clip=clip, guide_features=guide_features, **step_params)
 
@@ -98,8 +101,8 @@ def deepdream(filename, net, base_img, iter_n=10, octave_n=4, octave_scale=1.4, 
                 vis = vis*(255.0/np.percentile(vis, 99.98))
             #showarray(vis)
             if ((i > 10) and (i % 10 == 0)):
-                save_file(vis, end, i, octave, octave_scale, filename, model,guide)
-            print filename, octave, i, end, vis.shape
+                save_file(vis, end, i, octave, octave_scale, filename, model,guide_file)
+            print filename, octave, i, end, vis.shape, guide_file
             clear_output(wait=True)
 
         # extract details produced on the current octave
@@ -117,21 +120,20 @@ def objective_guide(dst, guide_features):
     dst.diff[0].reshape(ch,-1)[:] = y[:,A.argmax(1)] # select ones that match best
 
 
-def process2(net, frame, filename, model, guide):
-    #googlenet best
-    # layers = [
-    #         "inception_3b/5x5_reduce",
-    #         ]
-
-    #alexnet
-    # layers = ['conv3', 'conv4', 'conv5', 'pool5', 'conv1', 'norm1', 'pool1', 'conv2', 'norm2', 'pool2']
-
+def process2(net, frame, filename, model, guide, guide_file):
     layers = net.blobs.keys()
-    print layers
+    #print layers
 
-    for octave_n in [3]:
+    #googlenet best
+    layers = [
+            "inception_3b/5x5_reduce",
+            "inception_3b/5x5",
+            ]
+
+
+    for octave_n in [4]:
         for octave_scale in [2.5]:
-            for iterations in [100]:
+            for iterations in [50]:
                 for layer in layers:
                     if layer != "data":
                         if guide != None:
@@ -142,28 +144,27 @@ def process2(net, frame, filename, model, guide):
                             src.data[0] = preprocess(net, guide)
                             net.forward(end=layer)
                             guide_features = dst.data[0].copy()
-                            output = deepdream(filename, net, frame, iter_n=iterations, octave_n=octave_n, octave_scale=octave_scale, end=layer, model=model,objective=objective_guide,guide_features=guide_features,guide=True)
-                            save_file(output, layer, iterations, octave_n, octave_scale, "_____final_____" + filename, model, true)
+                            output = deepdream(filename, net, frame, iter_n=iterations, octave_n=octave_n, octave_scale=octave_scale, end=layer, model=model,objective=objective_guide,guide_features=guide_features,guide=guide, guide_file=guide_file)
+                            print("####################DONE GUIDE#############################")
+                            save_file(output, layer, iterations, octave_n, octave_scale, "_____final_____" + filename, model, guide_file)
                         else:
                             output = deepdream(filename, net, frame, iter_n=iterations, octave_n=octave_n, octave_scale=octave_scale, end=layer, model=model)
                             save_file(output, layer, iterations, octave_n, octave_scale, "_____final_____" + filename, model)
 
-def save_file(output, layer, iterations, octave_n, octave_scale, filename, model,guide=False):
-    name = ""
-    if guide == True:
-       name = "guided__"
-    name = name + model + "_" + layer.replace("/", "") + "_itr_" + str(iterations) + "_octs_"
+def save_file(output, layer, iterations, octave_n, octave_scale, filename, model,guide=""):
+    name = guide + "____" + model + "_" + layer.replace("/", "") + "_itr_" + str(iterations) + "_octs_"
     name2 = name + str(octave_n) + "_scl_" + str(octave_scale) + "_jt_"
     name3 = name2 + "32__nonlin" + filename
     PIL.Image.fromarray(np.uint8(output)).save("outputs/" + name3, dpi=(600,600))
 
 
-def start(filename):
+def start(filename, guide_file):
+    print("Starting async file processing")
     with open("settings.json") as json_file:
         json_data = json.load(json_file)
 
     try:
-        guide = np.float32(PIL.Image.open(os.getcwd() + "/guide.jpg"))
+        guide = np.float32(PIL.Image.open(os.getcwd() + "/guides/" + guide_file))
     except:
         guide = None
 
@@ -215,20 +216,29 @@ def start(filename):
     img = np.float32(img)
 
     frame = img
-    process2(net, frame, filename, model_name,guide)
+    print("++++++++++++++++++++++++++Processing File++++++++++++++++++++++++++")
+    process2(net, frame, filename, model_name,guide, guide_file)
+    shutil.move(os.getcwd() + "/inputs/" + guide_file, "guide_done/")
 
-    shutil.move(os.getcwd() + "/inputs/" + filename, "done/")
 
+###############################################################################
 count = 1
 for filename in os.listdir(os.getcwd() + "/inputs/"):
-    print filename
-    if filename == ".DS_Store" or filename == ".gitkeep":
-        continue
-    p = Process(target=start, args=(filename,))
-    p.start()
-    if count == 3:
-        print("BLOCKING.......")
-        p.join()
-        count = 1
-    else:
-        count = count + 1
+    for guide in os.listdir(os.getcwd() + "/guides/"):
+        if filename == ".DS_Store" or filename == ".gitkeep":
+            continue
+        if guide == ".DS_Store" or filename == ".gitkeep":
+            continue
+        print filename, guide
+        p = Process(target=start, args=(filename,guide,))
+        p.start()
+        if count == 1:
+            print("BLOCKING.......")
+            p.join()
+            count = 1
+        else:
+            count = count + 1
+    if filename != ".DS_Store" and filename != ".gitkeep":
+        shutil.move(os.getcwd() + "/inputs/" + filename, "done/")
+
+
